Several format changes for generated tweets
For replies, construct a header that lists the accounts being replied to, and the links to the tweet being replied to. Removes all reply-to "@username "s from the start of the body. 

Adds a '> ' quote prefix to all lines in the body.

Replaces "Originally on Twitter: " with the favicon as a scaled image.

Removes a case where it looks like video media url was being written unnecessarily to the output.